fix: resolve gcc warnings flagged by the ROS buildfarm (Humble/Jazzy) (#1412)
* fix: resolve gcc warnings flagged across the ROS buildfarm (Humble/Jazzy)

Addresses every warning category surfaced by the last Humble (Ubuntu
Jammy, gcc-11) and Jazzy (Ubuntu Noble, gcc-13) ROS buildfarm dev jobs:
-Wold-style-cast, -Wconversion, -Wsign-conversion, -Wfloat-conversion,
-Wunused-parameter, -Wunused-result, -Wdeprecated-declarations,
-Wdangling-else, -Wdangling-reference, -Wnon-virtual-dtor,
-Woverloaded-virtual, -Wsign-compare, -Wstringop-overread, -Wshadow,
-Wreturn-local-addr and -Wrange-loop-construct, spanning mrpt_bayes,
mrpt_containers, mrpt_core, mrpt_graphs, mrpt_graphslam, mrpt_hwdrivers,
mrpt_img, mrpt_maps, mrpt_math, mrpt_nav, mrpt_obs, mrpt_poses,
mrpt_serialization, mrpt_slam, mrpt_system and mrpt_viz.

Real bugs fixed along the way (not just silenced casts):

* CPose3DQuat::operator[](i) const returned a temporary `double` by
  value while its `const_iterator` bound it to a `const double&`,
  making every dereference of a const iterator dangling
  (-Wreturn-local-addr).
* CDirectedTree::Visitor and every one of its subclasses had virtual
  functions but a non-virtual destructor.
* CSerializable's schema-archive overloads of serializeTo/serializeFrom
  were silently hidden in any class using DEFINE_SERIALIZABLE, so they
  were unreachable without explicit qualification; added the missing
  `using` declarations.
* mrpt::Clock::time_point had no discoverable (ADL-reachable)
  operator<<, which made gtest's EXPECT_EQ/ASSERT_EQ fail to compile
  wherever it compared two timestamps on gcc-11/older libstdc++ (the
  actual Humble build error, not just a warning) -- this was the one
  case where mrpt::system::operator<<(TTimeStamp) already existed but
  lived in a namespace ADL never finds from outside mrpt::system; it is
  now `mrpt::operator<<`, with `mrpt::system` bringing it back in via a
  `using` declaration for existing qualified callers.
* Several TKFMethod-shaped unscoped enums (TKFMethod,
  TLaserSimulUncertaintyMethod, TMatrixTextFileFormat,
  TICPCovarianceMethod) had no fixed underlying type, so a test
  casting an intentionally-invalid value into them triggered
  "conversion is unspecified" rather than the well-defined reinterpret
  a fixed underlying type gives.

Everything else is either a widening/narrowing cast made explicit, an
unused parameter renamed to a comment, `else`-ambiguous one-liners
given braces, or a documented false-positive (e.g. a wrapper around a
cached-static-map accessor gcc's -Wdangling-reference still flags)
resolved by dropping the unnecessary named reference instead of
suppressing the warning.

Verified with an incremental colcon build of all 16 touched packages
under mrpt_common's warning set (-Wall -Wextra -Wshadow -Wtype-limits
-Wcast-align -Wparentheses -Wunused -Wpedantic -Wconversion
-Wsign-conversion -Wdouble-promotion -Woverloaded-virtual
-Wold-style-cast -Wnon-virtual-dtor): zero warnings left other than a
gcc-13/-O3 -Wmaybe-uninitialized false positive on small fixed-size
CMatrixDynamic construction that isn't present in either buildfarm log
and isn't touched by this change. All existing unit tests
(2000+ cases across the touched packages) still pass.


* fix: replace the Clock::time_point operator<< with a gtest PrintTo hook

The operator<<(Clock::time_point) added in the previous commit, found via
ADL because mrpt::Clock is in namespace mrpt, broke unrelated `os << x`
calls in any translation unit with several `using namespace mrpt::...;`
directives active at once (e.g. MonteCarloLocalization_App.cpp, which
has ~15 of them): overload resolution for `os << particles_count`
(a std::vector<int>) stopped finding mrpt::math's own
operator<<(ostream&, const std::vector<T>&) entirely -- this broke all
four CI compilers (gcc, clang, AppleClang, MSVC), all failing on the
exact same line.

It was also never the right fix for a second reason: mrpt::system
already had an operator<<(TTimeStamp) doing the same thing (TTimeStamp
being just an alias of Clock::time_point), so having both in scope at
once was flat-out ambiguous the moment `using namespace mrpt::system;`
and ADL both applied -- which is exactly the CObservation.cpp failure
this was meant to fix in the first place.

gtest's own recommended mechanism for exactly this situation --
teaching it to print a type it doesn't otherwise know how to -- is a
`PrintTo(value, ostream*)` overload, looked up via ADL *before* gtest
falls back to operator<<. It doesn't add anything to the type's normal
operator<< overload set, so it cannot participate in, or break, any
unrelated overload resolution. mrpt::system::operator<<(TTimeStamp) is
reverted to its original, unmodified form.

Verified with a clean incremental build of all 21 packages (the 16 from
the previous commit plus mrpt_libapps_gui/_cli, mrpt_gui, mrpt_apps_gui/_cli,
which pulled in MonteCarloLocalization_App.cpp): zero errors, zero
warnings. The four originally-affected unit test binaries
(mrpt_core, mrpt_system, mrpt_containers, mrpt_obs) still pass in full.


* TKFMethod enum class defined as uint8_t

+++ modules/mrpt_img/tests/CImage_SSEx_unittest.cpp
@@ -53,7 +53,8 @@ std::vector<uint8_t> makeGray(int w, int h)
   {
     for (int x = 0; x < w; x++)
     {
-      v[static_cast<size_t>(y) * w + x] = static_cast<uint8_t>((x + y * 7) % 256);
+      v[static_cast<size_t>(y) * static_cast<size_t>(w) + static_cast<size_t>(x)] =
+          static_cast<uint8_t>((x + y * 7) % 256);
     }
   }
   return v;
@@ -66,7 +67,8 @@ std::vector<uint8_t> makeColor(int w, int h)
   {
     for (int x = 0; x < w; x++)
     {
-      const size_t off = (static_cast<size_t>(y) * w + x) * 3;
+      const size_t off =
+          (static_cast<size_t>(y) * static_cast<size_t>(w) + static_cast<size_t>(x)) * 3;
       v[off + 0] = static_cast<uint8_t>((x * 3 + y) % 256);
       v[off + 1] = static_cast<uint8_t>((x * 5 + y * 2) % 256);
       v[off + 2] = static_cast<uint8_t>((x * 7 + y * 3) % 256);
@@ -83,7 +85,7 @@ TEST(CImage_SSEx, ScaleHalf1c8u_MultipleOf16)
   const int w = 32;
   const int h = 4;
   const auto in = makeGray(w, h);
-  std::vector<uint8_t> out(static_cast<size_t>(w / 2) * (h / 2), 0);
+  std::vector<uint8_t> out(static_cast<size_t>(w / 2) * static_cast<size_t>(h / 2), 0);
 
   image_SSE2_scale_half_1c8u(in.data(), out.data(), w, h, w, w / 2);
 
@@ -92,7 +94,8 @@ TEST(CImage_SSEx, ScaleHalf1c8u_MultipleOf16)
     for (int ox = 0; ox < w / 2; ox++)
     {
       EXPECT_EQ(
-          out[static_cast<size_t>(oy) * (w / 2) + ox], in[static_cast<size_t>(2 * oy) * w + 2 * ox])
+          out[static_cast<size_t>(oy) * static_cast<size_t>(w / 2) + static_cast<size_t>(ox)],
+          in[static_cast<size_t>(2 * oy) * static_cast<size_t>(w) + static_cast<size_t>(2 * ox)])
           << "at (" << ox << "," << oy << ")";
     }
   }
@@ -104,7 +107,7 @@ TEST(CImage_SSEx, ScaleHalf1c8u_WithRemainder)
   const int w = 20;
   const int h = 2;
   const auto in = makeGray(w, h);
-  std::vector<uint8_t> out(static_cast<size_t>(w / 2) * (h / 2), 0);
+  std::vector<uint8_t> out(static_cast<size_t>(w / 2) * static_cast<size_t>(h / 2), 0);
 
   image_SSE2_scale_half_1c8u(in.data(), out.data(), w, h, w, w / 2);
 
@@ -119,8 +122,8 @@ TEST(CImage_SSEx, ScaleHalfSmooth1c8u_UniformBlock)
 {
   const int w = 32;
   const int h = 4;
-  std::vector<uint8_t> in(static_cast<size_t>(w) * h, 100);
-  std::vector<uint8_t> out(static_cast<size_t>(w / 2) * (h / 2), 0);
+  std::vector<uint8_t> in(static_cast<size_t>(w) * static_cast<size_t>(h), 100);
+  std::vector<uint8_t> out(static_cast<size_t>(w / 2) * static_cast<size_t>(h / 2), 0);
 
   image_SSE2_scale_half_smooth_1c8u(in.data(), out.data(), w, h, w, w / 2);
 
@@ -135,7 +138,7 @@ TEST(CImage_SSEx, ScaleHalfSmooth1c8u_WithRemainderStaysInRange)
   const int w = 20;
   const int h = 2;
   const auto in = makeGray(w, h);
-  std::vector<uint8_t> out(static_cast<size_t>(w / 2) * (h / 2), 0);
+  std::vector<uint8_t> out(static_cast<size_t>(w / 2) * static_cast<size_t>(h / 2), 0);
 
   image_SSE2_scale_half_smooth_1c8u(in.data(), out.data(), w, h, w, w / 2);
 
@@ -144,10 +147,10 @@ TEST(CImage_SSEx, ScaleHalfSmooth1c8u_WithRemainderStaysInRange)
   // source block.
   for (int ox = 0; ox < w / 2; ox++)
   {
-    const int a = in[2 * ox];
-    const int b = in[2 * ox + 1];
-    const int c = in[w + 2 * ox];
-    const int d = in[w + 2 * ox + 1];
+    const int a = in[static_cast<size_t>(2 * ox)];
+    const int b = in[static_cast<size_t>(2 * ox + 1)];
+    const int c = in[static_cast<size_t>(w + 2 * ox)];
+    const int d = in[static_cast<size_t>(w + 2 * ox + 1)];
     const int lo = std::min({a, b, c, d});
     const int hi = std::max({a, b, c, d});
     EXPECT_GE(out[static_cast<size_t>(ox)], lo);
@@ -171,11 +174,14 @@ TEST(CImage_SSEx, ScaleHalf3c8u_MultipleOf16)
   {
     for (int ox = 0; ox < w / 2; ox++)
     {
-      const size_t outOff = (static_cast<size_t>(oy) * (w / 2) + ox) * 3;
-      const size_t inOff = (static_cast<size_t>(2 * oy) * w + 2 * ox) * 3;
+      const size_t outOff =
+          (static_cast<size_t>(oy) * static_cast<size_t>(w / 2) + static_cast<size_t>(ox)) * 3;
+      const size_t inOff =
+          (static_cast<size_t>(2 * oy) * static_cast<size_t>(w) + static_cast<size_t>(2 * ox)) * 3;
       for (int c = 0; c < 3; c++)
       {
-        EXPECT_EQ(out[outOff + c], in[inOff + c]) << "at (" << ox << "," << oy << ") c=" << c;
+        EXPECT_EQ(out[outOff + static_cast<size_t>(c)], in[inOff + static_cast<size_t>(c)])
+            << "at (" << ox << "," << oy << ") c=" << c;
       }
     }
   }
@@ -197,7 +203,8 @@ TEST(CImage_SSEx, ScaleHalf3c8u_WithRemainder)
     const size_t inOff = static_cast<size_t>(2 * ox) * 3;
     for (int c = 0; c < 3; c++)
     {
-      EXPECT_EQ(out[outOff + c], in[inOff + c]) << "at ox=" << ox << " c=" << c;
+      EXPECT_EQ(out[outOff + static_cast<size_t>(c)], in[inOff + static_cast<size_t>(c)])
+          << "at ox=" << ox << " c=" << c;
     }
   }
 }
@@ -232,7 +239,7 @@ void checkUniformGray(
     const char* label)
 {
   const auto in = makeUniformRgb(w, h, r, g, b);
-  std::vector<uint8_t> out(static_cast<size_t>(w) * h, 0);
+  std::vector<uint8_t> out(static_cast<size_t>(w) * static_cast<size_t>(h), 0);
   fn(in.data(), out.data(), w, h, static_cast<size_t>(w) * 3, static_cast<size_t>(w));
 
   for (auto v : out)
